+++ test/libsolidity/syntaxTests/controlFlow/storageReturn/tuple_fine.sol
@@ -8,5 +8,8 @@ contract C {
         uint a;
         (c, a) = f();
     }
+    function h() internal view returns (S storage, S storage) {
+        return (s,s);
+    }
 }
 // ----